Make all pages server pages, convert login and registration to form components
